Added environment + package dependencies
Basic files to help with package distribution:
Virtual (conda) environment dependencies handled in the
environment.yml file.
Package dependencies to help build the actual package are handled
in the setup.cfg file, along with the dummy (but necessary)
setup.py file.
The requirements.txt is empty for now, but it is there to list any
concrete dependencies for the package that we may add later.