Added ErosionRate to SoilMass ODE in Soil Module
ErosionRate is the rate of soil mass loss. The calculation of this
in the python code differs (modification) from that of the Stella
code, as I was unsure that the Stella code was using appropriate
equations and units. The SoilMass ODE works now as expected and I
have also ensured the coupled PlantSoilModel works with this update