Added discrete planting flux, day-of-year forcing
The ordinal simulation day is now included in the forcing data
so that discrete management events (e.g. planting on day-of-year
30) can occur. Before this, with the scipy solver, you cannot
reference the time of the simulation to undertake this kind of
calculation. This same approach can be used to define multiple
different discrete events that depend on a reference to calendar
date or time.