Testing new module for Management
Separating out the Management practices/processes into its own
module. At the moment, this is only being passed into the Plant
module and being applied to the PhBioPlanting and NPhBioPlanting
fluxes. It is working so far. Next steps will fully integrate the
Management module into the Plant and Soil modules.