Corrected minor bugs, added notebook state change test
Corrected a few minor bugs, including: (i) call to
calculate_Transdown method and use of np.vectorize with conditional
function required defining otypes argument, otherwise it returned
integer instead of float values (it only did this when calling
the function directly in a diagnostic way, not when the solver was
running – related to some issue with np.float vs np.array of size
zero), (ii) Callable input arguments to solver method, (iii)
shifted state variable Bio_time to the front of argument list for
consistency.

In notebook example, included calculation of plant state variables
from the diagnostic fluxes to cross-check with the solver states.
Note that these differ at the moment because of differences in
instantaneous flux calculations and integration approaches. I will
resolve this in a later commit.