Replaced solar forcing in Plant with Site and date inputs
Solar-related variables used in the Plant module (e.g. sunset,
sunrise, daylength) are now calculated within the module instead of
being passed in as forcing data inputs. Instead, we now define a
Site module, which contains the parameters and methods needed to
calculate sunset, sunrise, and daylength with date (year, ordinal
day-of-year), latitude, longitude and timezone.

This reduces the demand on a user that just wants to run the Plant
module. They do not need to calculate these forcing data to provide
it as input. Instead, just define the ordinal day-of-year and year
of your simulation, along with the site/location-specific
information (latitude, longitude, timezone) and all of the solar
variables will be calculated internally.

Note: As part of this I also had to include a new interpolation
method for the forcing variables "year" and "doy". This
interpolation method is called piecewise constant (function
interp_nearest_lower_neighbour). At transition points across years
e.g. t(DOY=365, year=2018) to t(DOY=1, year=2019) a linear
interpolation method will return incorrect values for both DOY and
year. A piecewise constant interpolation is needed.