Removed dynamic calc of canopy nlayers
The canopy layers module included methods that were able to
dynamically determine an appropriate number of discrete layers
based on the canopy height and leaf area index. This implementation
was fine when running the model with one time-step. However, when
trying the do diagnostic calculations where e.g. LAI is passed in
as an array of values, this dynamic approach fails as you cant set
a class attribute as an array. I have thus removed this approach
so you must initialise the canopy layers module with nlevmlcan
explicitly and it will remain fixed at that number.