Swapped alternative leafgasexchange into CanopyGasExchange
Replaced the leafgasexchange module in the canopy gas exchange
notebook example with the alternative leafgasexchange2 module.
This is based on Duursma's plantecophys package. This uses the
Farquhar photosynthetic biochemical model and the Medlyn 2011
stomatal conductance model. It provides more stable results across
more conditions, as it implements the optimal stomatal conductance
model using a quadratic, rather than with a iterative loop.