Fix on climate forcing interp method
The interp_forcing with kind="pconst" method, which uses the
interp_nearest_lower_neighbour method, was returning outputs in
the wrong form (the values were correct). This fix ensures that
when the interpolation is called with a scalar input (e.g. during
ODE solving), it returns a scalar numpy array (size=1, shape=()),
and when called with an array input, it returns an array of the
appropriate shape.