Design new numerical solver
Designed and tested a new approach to the numerical solver for
solving the ordinary differential equations in the model. This new
approach is much more generic and "stand-alone", as it does not
require extensive modifications when using it for different models.
It includes two solver options: scipy's solve_ivp or an explicit
Eulers method. Handling of the model Callable, input forcing and
zero-crossing events (where a state is modified mid-run) are all
generic and adaptable.