Removed GDD reset hack
Removed the hack that was used to reset the growing degree days
state variable in the plant model. This hack detected when the
plantingDay occurred, then removed a large "flux" from the Bio_time
state, triggering a zero-crossing event in the solver. With the new
solve we handle this event explicitly.