Dev root zone weighting for water uptake
I have added a relatively simple approach to weight the root zone
soil-to-root hydraulic conductivity and soil water potential so
that the single, weighted values can be implemented in the function
that solves for the leaf water potential. This approach assumes that
the relative water uptake per layer is dependent on the soil-to-root
hydraulic conductivity (K_sr). If K_sr is relatively higher in a
given layer, that layer will contribute relatively more to water
uptake