Update plant_1000 for multi-layer soil
Incorporated the multi-layer soil module into the plant_1000 model.
It now takes multi-layer soil moisture as a forcing input and
passes this to the plantcarbonwater.py module. Also fixed up the
plantcarbonwater.py module to avoid use of a time dimension in the
soil moisture input, as this is inconsistent with how this module
is used elsewhere and causes headaches when combined with
plant_1000 and the ODE solver.