Added wheat grain production to plant_1000
Added class methods to plant_1000 that describe grain production
in wheat. This required addition of an extra state variable, which
is simply the stem pool but only updated during a specific growth/
developmental phase, which provides the basis of the grain
production model. In future, this should probably be abstracted
into a separate class for grain production, or modify plant_1000
to be only for wheat.