Fixed inconsistencies in plant respiration
Fixes: (1) Maintenance and growth respiration now modelled separately,
rather than with a simple carbon-use efficiency parameter in
plant_1000. (2) Maintenance respiration of leaf biomass is assumed
to equal the total canopy leaf mitochondrial respiration, as
determined in the leaf gas exchange module. (3) Maintenance
respiration of root biomass is modelled as a function of a maintenance
respiration coefficient, temperature and the amount of root biomass.
(4) Growth respiration is calculated as a fraction of available
assimilates after maintenance is accounted for
i.e. Rg=alpha_Rg*(GPP – Rm).