Updated germination-emergence process
Implemented a sowing-germination-emergence process whereby we no
longer need to initialise the model with leaf, stem or root biomass.
Instead, we can initialise the states all as zero, then set a
sowingDay and sowingRate, which is divided into leaf/stem/root
pools at the time of emergence. Emergence is determined by the
growing degree days and a sowing depth modifier.