Fixed two bugs in plantcarbonwater
The first bug was just due to the recent commit where SAI was added
as an input variable to the calculate function - it wasn't being
passed onto the method leaf_water_potential_solve. The second bug
was that root water potential was being calculated with the wrong
transpiration rate (it was using total canopy transpiration instead
of the leaf-area specific transpiration rate) which made it appear
to be more negative than leaf water potential (which doesn't make
physical sense in this model that relies on the water potential
gradient between the roots and leaves).