Added feature to store diagnostic outputs with solver
Added a feature to the ODEModelSolver that can store model
ODE calculator outputs (diagnostic or intermediate values) that
control the dynamics of the states e.g. fluxes or dynamic plant
properties. These were previously not being saved during the ODE
solver runs, and had to be diagnosed afterwards with a brand new
set of simulations, which was inefficient and prone to errors. Here,
diagnostic outputs are written from the calculate method for the
dydt function (as implemented in plant_1000 in this commit) and
stored as class attributes in the ODEModelSolver.