Updated plant_1000 notebook example for diagnostic outputs
In the previous commit I added a feature to store diagnostic
outputs with solver, rather than have to calculate them after the
solver has solved for the state variables. I have updated the
example notebook for plant_1000, including all of the scalar
diagnostics and figures, given this update in model output.