Split grain production into canola or wheat options
The process of grain production is considered uniquely for canola
and wheat crop types. So, I have included a cropType choice in the
management module that determines which grain production module
is called in plant_1000. All other processes are considered the
same (although parameterization would have to be modified between
crop types).