Corrected bug in units for minimum stomatal conductance
The minimum stomatal conductance parameter, g0, used in
leafgasexchange2 is typically expressed in native units of
conductance to water vapour (mol H2O m-2 s-1). The coupled
model requires it to be in conductance to CO2. We did this
conversion in the model but did not pass it to the sub-routines.