Fix on plant_1000 for phys parameter downregulation
If we don't return the parameters to their original values after
the downregulation phase, the initialized model class will keep
the adjusted parameter values which were not what was oiginally
defined. This overall approach for adjusting class parameters on
the fly is poor modeling practice and will need fixing/updating
in future.