Feature to active a process in multiple dev phases
The processes such as stem remobilization to grain and allocation
of NPP to grain were only activated in a single developmental phase.
Here, I have added a feature that allows multiple dev phases to
be listed and these processes to be activated in any of them. This
requires to plant_1000 attributes "remob_phase" and "grainfill_phase"
to be a list of strings.