Another fix on plant_1000 phys parameter downregulation
Need to restore the downregulated physiological parameters
(class attributes) at the end of each calculate method, otherwise
there can be issues when re-running the model a second time, as the
attributes are not their original value. Again, this overall
approach for adjusting class parameters on the fly is poor modeling
practice and will need fixing/updating in future.