Fixed error in LAI, Cleaf, W_L relationship
Because of the smoothed minimum function used to calculate LAI,
the optimal allocation scheme could continue adding carbon to the
leaf pool even though the LAI derived from Cleaf and SLA (or W_L
and SLA) exceeded the maxLAI parameter. This broke the assumed
relationship between LAI and SLA. Now, the smoothed minimum
function has been removed, which influences the dynamics of the
carbon allocation scheme. So, in addition, if the allocation
scheme sets leaf and root allocation coefficients to zero, we
instead direct that allocation fraction to stem, to ensure carbon
is conserved.