Added empirical function for leaf temperature
Previously, the leaf temperatures used in the calculation of the
plant carbon and water dynamics (plantcarbonwater) was simply set
to the daily average air temperature. Here, we have derived a
simple empirical relationship from a field site that determines
surface skin temperature based on air temperature and downwelling
shortwave radiation.