Added try-except for Psi_l bisection method
To solve the plant water balance and calculate the leaf water
potential, we use the bisection method. In some edge cases when
conditions are rather extreme (e.g. very low soil moisture or LAI)
this numerical method does not work, but we want the model to
continue running. I have added a try-except block for handling
this, setting Psi_l to a value that will cause 99% stomatal down-
regulation (correspoding to dessication).