Fix phys parameter downregulation phase dependence
The downregulated physiological parameters (class attributes) are
restored to the original values at the end of the defined
downreg_phase. However, this can occur prior to a harvest event or
restart of new growing season (e.g. maturity phase can end before
harvest), in which case the code would restore these parameters to
their original values - this behaviour is not desired as we want
a matured plant to stay mature. This has been fixed.