New post-processing time indexing support
New method in the Climate (Site) module that helps to determine
time indexing of model outputs to support post-processing analysis.
This includes finding the time index of sowing, maturity, harvest and
phase transition points within the growing season you want to
analyse. This makes the post-processing easier and translatable to
other implementations of the model.