Updated optimal alloc gradient methods
The plant optimal carbon allocation module calculates the gradient
of the net carbon profit with respect to (i) the leaf biomass pool,
and (ii) the root biomass pool. The method for doing with was
at times problematic, leading to sudden jumps in the gradient and
hence calculated allocation coefficients. Here, I have updated the
module to include three options for determining the gradient, with
different step sizes to make it more flexible. N.B. This fix
changes the outputs of the plant_1000 runs.