Added funcs to utils to recreate modules from netcdf file
I have added some helper scripts that take a long string input of
any DAESIM2 module (e.g. PlantGrowthPhases) and convert it into an
the actual initialized module. In each netcdf output file there is
an attribute DAESIMcalculator, which is a string of the python
class, documenting the model that was used to generate those
outputs. These new scripts in utils allow us to interpret and use
the netcdf file attribute DAESIMcalculator that is written to the
outputs. Important for reproducibility.