Updated management sowing/harvest attribute type
The attributes of the Management module that define sowing and
harvest dates (sowingDays, sowingYears, harvestDays, harvestYears)
could be initiliazed as integers which can cause problems in some
daesim2_analysis package methods. These should always be initiliased
as lists, so here, I update the class code to allow the attributes
to be set with integers but ensure the class assigns these to a
list by default (using a __post_init__ method). This means these
attributes are by default an empty list (and not None like they
were before) so I also had to update conditional statements in
the plant calculators to reflect this change.